Fix -Wformat errors found in ircd-ratbox.
We cannot use -Wformat meaningfully but ircd-ratbox trunk can.

+++ modules/core/m_join.c
@@ -363,7 +363,8 @@ m_join(struct Client *client_p, struct Client *source_p, int parc, const char *p
 
 			sendto_one(source_p, form_str(RPL_TOPICWHOTIME),
 				   me.name, source_p->name, chptr->chname,
-				   chptr->topic_info, chptr->topic_time);
+				   chptr->topic_info,
+				   (unsigned long)chptr->topic_time);
 		}
 
 		channel_member_names(chptr, source_p, 1);
