+++ modules/m_topic.c
@@ -163,7 +163,8 @@ m_topic(struct Client *client_p, struct Client *source_p, int parc, const char *
 
 			sendto_one(source_p, form_str(RPL_TOPICWHOTIME),
 					me.name, source_p->name, chptr->chname,
-					chptr->topic_info, chptr->topic_time);
+					chptr->topic_info,
+					(unsigned long)chptr->topic_time);
 		}
 	}
